requirements.txt
# Core Dependencies
python-dotenv==1.0.0
anyio>=3.7.1

# FastAPI Stack
typing-extensions>=4.8.0
pydantic>=2.0.0,<3.0.0
starlette>=0.27.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
sse-starlette>=1.6.5
httpx>=0.24.0

# Database
psycopg2-binary>=2.9.9
supabase>=2.0.0  # Updated for Python 3.12 compatibility

# LangChain Ecosystem
langchain>=0.0.335
langchain-core>=0.1.17
langsmith>=0.0.83
openai>=1.12.0

# Data Processing
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
PyPDF2>=3.0.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
tenacity>=8.2.3
tqdm>=4.66.1
python-dateutil>=2.8.2
orjson>=3.9.10
websockets>=12.0
aiofiles>=23.2.1

# Scheduling
apscheduler>=3.10.1
schedule>=1.2.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Installation Instructions
# 1. Create and activate a new virtual environment:
#    python -m venv venv
#    venv\Scripts\activate  # Windows
#
# 2. Install requirements:
#    pip install --upgrade pip
#    pip install -r requirements.txt
#
# Note: This setup uses Pydantic v1. Do not install packages that require Pydantic v2.
